+++ sdk/openai/openai/sources/customizations/OpenAIClient.ts
@@ -2,25 +2,27 @@
 // Licensed under the MIT license.
 
 import { TokenCredential, KeyCredential, isTokenCredential } from "@azure/core-auth";
-import {
-  GetAzureBatchImageGenerationOperationStatusOptions,
-  GetCompletionsOptions,
-  GetEmbeddingsOptions,
-} from "../generated/src/models/options.js";
+import { GetCompletionsOptions, GetEmbeddingsOptions } from "../generated/src/models/options.js";
 import { OpenAIClientOptions } from "../generated/src/index.js";
-import { getChatCompletions, listChatCompletions, listCompletions } from "./api/operations.js";
-import { ChatMessage, Completions, Embeddings } from "../generated/src/models/models.js";
 import {
-  beginAzureBatchImageGeneration,
-  getAzureBatchImageGenerationOperationStatus,
-  getCompletions,
-  getEmbeddings,
-} from "../generated/src/api/operations.js";
-import { ChatCompletions, ImageGenerationResponse } from "./models/models.js";
+  getChatCompletions,
+  getImages,
+  listChatCompletions,
+  listCompletions,
+} from "./api/operations.js";
+import {
+  ChatMessage,
+  Completions,
+  Embeddings,
+  ImageGenerations,
+} from "../generated/src/models/models.js";
+import { getCompletions, getEmbeddings } from "../generated/src/api/operations.js";
+import { ChatCompletions } from "./models/models.js";
 import { OpenAIContext } from "../generated/src/rest/index.js";
 import { createOpenAI } from "../generated/src/api/OpenAIContext.js";
 import { GetChatCompletionsOptions } from "./api/models.js";
 import { ImageGenerationOptions } from "./models/options.js";
+import { PipelinePolicy } from "@azure/core-rest-pipeline";
 
 function createOpenAIEndpoint(version: number): string {
   return `https://api.openai.com/v${version}`;
@@ -129,17 +131,7 @@ export class OpenAIClient {
               ...(opts.additionalPolicies ?? []),
               {
                 position: "perCall",
-                policy: {
-                  name: "openAiEndpoint",
-                  sendRequest: (request, next) => {
-                    const obj = new URL(request.url);
-                    const parts = obj.pathname.split("/");
-                    obj.pathname = `/${parts[1]}/${parts.slice(5).join("/")}`;
-                    obj.searchParams.delete("api-version");
-                    request.url = obj.toString();
-                    return next(request);
-                  },
-                },
+                policy: getPolicy(),
               },
             ],
           }),
@@ -232,24 +224,6 @@ export class OpenAIClient {
     return listChatCompletions(this._client, messages, deploymentName, options);
   }
 
-  /** Returns the status of the images operation */
-  getAzureBatchImageGenerationOperationStatus(
-    operationId: string,
-    options: GetAzureBatchImageGenerationOperationStatusOptions = {
-      requestOptions: {},
-    }
-  ): Promise<ImageGenerationResponse> {
-    return getAzureBatchImageGenerationOperationStatus(this._client, operationId, options);
-  }
-
-  /** Starts the generation of a batch of images from a text caption */
-  beginAzureBatchImageGeneration(
-    prompt: string,
-    options: ImageGenerationOptions = { requestOptions: {} }
-  ): Promise<ImageGenerationResponse> {
-    return beginAzureBatchImageGeneration(this._client, prompt, options);
-  }
-
   /**
    * Starts the generation of a batch of images from a text caption
    * @param prompt - The prompt to use for this request.
@@ -259,7 +233,35 @@ export class OpenAIClient {
   getImages(
     prompt: string,
     options: ImageGenerationOptions = { requestOptions: {} }
-  ): Promise<ImageGenerationResponse> {
-    return beginAzureBatchImageGeneration(this._client, prompt, options);
+  ): Promise<ImageGenerations> {
+    return getImages(this._client, prompt, options);
   }
+}
+
+function getPolicy(): PipelinePolicy {
+  const policy: PipelinePolicy = {
+    name: "openAiEndpoint",
+    sendRequest: (request, next) => {
+      const obj = new URL(request.url);
+      const parts = obj.pathname.split("/");
+      switch (parts[parts.length - 1]) {
+        case "completions":
+          if (parts[parts.length - 2] === "chat") {
+            obj.pathname = `/${parts[1]}/chat/completions`;
+          } else {
+            obj.pathname = `/${parts[1]}/completions`;
+          }
+          break;
+        case "embeddings":
+          obj.pathname = `/${parts[1]}/embeddings`;
+          break;
+        case "generations:submit":
+          obj.pathname = `/${parts[1]}/images/generations`;
+      }
+      obj.searchParams.delete("api-version");
+      request.url = obj.toString();
+      return next(request);
+    },
+  };
+  return policy;
 }