+++ sdk/openai/openai/sources/customizations/api/operations.ts
@@ -1,24 +1,37 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ChatMessage, Completions } from "../../generated/src/models/models.js";
+import {
+  ChatMessage,
+  Completions,
+  ImageGenerations,
+  ImageLocation,
+} from "../../generated/src/models/models.js";
 import { GetCompletionsOptions } from "../../generated/src/models/options.js";
 import {
   _getCompletionsSend,
   _getChatCompletionsSend,
   _getChatCompletionsWithAzureExtensionsSend,
+  _beginAzureBatchImageGenerationSend,
 } from "../../generated/src/api/operations.js";
 import { getOaiSSEs } from "./oaiSse.js";
 import {
+  BeginAzureBatchImageGeneration202Response,
+  BeginAzureBatchImageGenerationDefaultResponse,
+  BeginAzureBatchImageGenerationLogicalResponse,
   OpenAIContext as Client,
   GetChatCompletionsWithAzureExtensions200Response,
   GetChatCompletionsWithAzureExtensionsDefaultResponse,
+  ImageGenerationsOutput,
+  ImagePayloadOutput,
+  getLongRunningPoller,
   isUnexpected,
 } from "../../generated/src/rest/index.js";
 import { StreamableMethod } from "@azure-rest/core-client";
 import { ChatCompletions } from "../models/models.js";
 import { getChatCompletionsResult, getCompletionsResult } from "./deserializers.js";
 import { GetChatCompletionsOptions } from "./models.js";
+import { ImageGenerationOptions } from "../models/options.js";
 
 export function listCompletions(
   context: Client,
@@ -33,6 +46,64 @@ export function listCompletions(
   return getOaiSSEs(response, getCompletionsResult);
 }
 
+export async function getImages(
+  context: Client,
+  prompt: string,
+  options: ImageGenerationOptions = { requestOptions: {} }
+): Promise<ImageGenerations> {
+  const response = await _beginAzureBatchImageGenerationSend(context, prompt, options);
+
+  if (isUnexpected(response)) {
+    // Check for response from OpenAI
+    const body = response.body as unknown as ImageGenerations;
+    if (body.created && body.data) {
+      return body;
+    }
+    throw response.body.error;
+  }
+
+  if (response.status === "202") {
+    const poller = await getLongRunningPoller(
+      context,
+      response as BeginAzureBatchImageGeneration202Response
+    );
+    const result = await poller.pollUntilDone();
+    return getImageResultsDeserialize(result);
+  } else {
+    return getImageResultsDeserialize(response);
+  }
+}
+
+function convertResultTypes({ created, data }: ImageGenerationsOutput): ImageGenerations {
+  if (typeof (data[0] as ImageLocation).url === "string") {
+    return {
+      created: new Date(created),
+      data: data as ImageLocation[],
+    };
+  } else {
+    return {
+      created: new Date(created),
+      data: data.map((item) => {
+        return {
+          base64Data: (item as ImagePayloadOutput).b64_json,
+        };
+      }),
+    };
+  }
+}
+
+function getImageResultsDeserialize(
+  response:
+    | BeginAzureBatchImageGeneration202Response
+    | BeginAzureBatchImageGenerationDefaultResponse
+    | BeginAzureBatchImageGenerationLogicalResponse
+): ImageGenerations {
+  if (isUnexpected(response) || !response.body.result) {
+    throw response.body.error;
+  }
+  const result = response.body.result;
+  return convertResultTypes(result);
+}
 function _getChatCompletionsSendX(
   context: Client,
   messages: ChatMessage[],