+++ sdk/openai/openai/CHANGELOG.md
@@ -13,6 +13,9 @@
 
 ### Breaking Changes
 
+- Remove `beginAzureBatchImageGeneration` and `getAzureBatchImageGenerationOperationStatus` methods. 
+- `getImages` has been updated to return the image URLs/paylods directly, rather than requiring the user to call `getAzureBatchImageGenerationOperationStatus` to retrieve them.
+
 ### Bugs Fixed
 
 ### Other Changes